refactor: 용도에 맞게 domain내 class들의 package 구분
