+++ src/main/java/blackjack/domain/cardGenerator/CardGenerator.java
@@ -1,6 +1,6 @@
-package blackjack.domain.CardGenerator;
+package blackjack.domain.cardGenerator;
 
-import blackjack.domain.PlayingCard;
+import blackjack.domain.card.PlayingCard;
 import java.util.Stack;
 
 public interface CardGenerator {
