+++ src/test/java/blackjack/domain/BlackJackResultTest.java
@@ -2,6 +2,9 @@ package blackjack.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import blackjack.domain.player.Dealer;
+import blackjack.domain.player.Gambler;
+import blackjack.domain.player.Player;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Map;
